Project8_Dall_e_image_generation/requirements.txt
openai==1.63.2
python-dotenv
flask